+++ CHANGELOG.md
@@ -69,3 +69,5 @@
 ## v1.3.9
 * Added param `customOptions` to the `getElevationValues` method, allowing to change the quality/resolution at each reading
 * Added helper methods `getMaxTilesResolution` and `getCurrentViewResolution`
+* Added event `load`
+* Improved `readFromImage` class to avoid some retiterated function calls on loops